Replace vector math classes with the Eigen::Matrix class
This replaces the StaticVector, StaticMatrix, GenericVector,
and GenericPoint classes with the Eigen::Matrix class. The
Point3 and Vector3 typedefs have also been replaced with
typedefs for Eigen::Matrix.

+++ examples/cube-viewer/cubeviewerexample.cpp
@@ -130,7 +130,7 @@ void CubeViewerExample::openFile(const QString &fileName)
     m_positiveScalarField = readVolumeData(fileName);
     if(m_positiveScalarField){
         m_positiveSurfaceItem->setScalarField(m_positiveScalarField);
-        m_positiveSurfaceItem->setPosition(m_positiveScalarField->origin());
+        m_positiveSurfaceItem->setPosition(m_positiveScalarField->origin().cast<float>());
 
         std::vector<chemkit::Float> values = m_positiveScalarField->data();
         for(unsigned int i = 0; i < values.size(); i++){
@@ -140,7 +140,7 @@ void CubeViewerExample::openFile(const QString &fileName)
                                                          m_positiveScalarField->cellDimensions(),
                                                          values);
         m_negativeSurfaceItem->setScalarField(m_negativeScalarField);
-        m_negativeSurfaceItem->setPosition(m_positiveScalarField->origin());
+        m_negativeSurfaceItem->setPosition(m_positiveScalarField->origin().cast<float>());
     }
 
     // update the view
